Refactor Member and ContainerTypeCode to use common applied annotations API (#107)
* Refs #19968: refactor Member and ContainerTypeCode to use common applied member annotations API


* Refs #19968: apply review suggestions


---------
