+++ src/main/java/com/eprosima/idl/parser/typecode/CollectionElement.java
@@ -1,0 +1,41 @@
+// Copyright 2023 Proyectos y Sistemas de Mantenimiento SL (eProsima).
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.eprosima.idl.parser.typecode;
+
+public class CollectionElement extends MemberAppliedAnnotations
+{
+    public CollectionElement()
+    {
+        super();
+    }
+
+    public CollectionElement(TypeCode typecode)
+    {
+        super();
+        typecode_ = typecode;
+    }
+
+    public TypeCode getTypecode()
+    {
+        return typecode_;
+    }
+
+    public void setTypecode(TypeCode typecode)
+    {
+        typecode_ = typecode;
+    }
+
+    private TypeCode typecode_ = null;
+}
