+++ src/main/java/com/eprosima/idl/parser/typecode/Member.java
@@ -14,27 +14,18 @@
 
 package com.eprosima.idl.parser.typecode;
 
-import com.eprosima.idl.parser.exception.RuntimeGenerationException;
-import com.eprosima.idl.parser.tree.Annotation;
-import com.eprosima.idl.parser.tree.Notebook;
-import com.eprosima.idl.context.Context;
-
-import java.util.Map;
-import java.util.HashMap;
-import java.util.Collection;
-
-public class Member implements Notebook
+public class Member extends MemberAppliedAnnotations
 {
     public Member()
     {
-        m_annotations = new HashMap<String, Annotation>();
+        super();
     }
 
     public Member(TypeCode typecode, String name)
     {
+        super();
         m_typecode = typecode;
         m_name = name;
-        m_annotations = new HashMap<String, Annotation>();
     }
 
     public String getName()
@@ -73,140 +64,6 @@ public class Member implements Notebook
          m_typecode = typecode;
     }
 
-    @Override
-    public void addAnnotation(Context ctx, Annotation annotation)
-    {
-        if(annotation != null)
-        {
-            m_annotations.put(annotation.getName(), annotation);
-        }
-    }
-
-    @Override
-    public Map<String, Annotation> getAnnotations()
-    {
-        return m_annotations;
-    }
-
-    public Collection<Annotation> getAnnotationList()
-    {
-        return m_annotations.values();
-    }
-
-    public boolean isAnnotationOptional()
-    {
-        Annotation ann = m_annotations.get("optional");
-        if (ann != null)
-        {
-            return ann.getValue().toUpperCase().equals("TRUE");
-        }
-        return false;
-    }
-
-    public boolean isAnnotationExternal()
-    {
-        Annotation ann = m_annotations.get("external");
-        if (ann != null)
-        {
-            return ann.getValue().toUpperCase().equals("TRUE");
-        }
-        return false;
-    }
-
-    public boolean isAnnotationMustUnderstand()
-    {
-        Annotation ann = m_annotations.get("must_understand");
-        if (ann != null)
-        {
-            return ann.getValue().toUpperCase().equals("TRUE");
-        }
-        return false;
-    }
-
-    public boolean isAnnotationNonSerialized()
-    {
-        Annotation ann = m_annotations.get("non_serialized");
-        if (ann != null)
-        {
-            return ann.getValue().toUpperCase().equals("TRUE");
-        }
-        return false;
-    }
-
-    public boolean isAnnotationKey()
-    {
-        Annotation ann = m_annotations.get("key");
-        if (ann == null)
-        {
-            ann = m_annotations.get("Key"); // Try old way
-        }
-        if (ann != null)
-        {
-            return ann.getValue().toUpperCase().equals("TRUE");
-        }
-        return false;
-    }
-
-    public Short getAnnotationBitBound()
-    {
-        Annotation ann = m_annotations.get("bit_bound");
-        if (ann != null)
-        {
-            String value = ann.getValue();
-            if (value.equals("-1"))
-            {
-                return null;
-            }
-            return Short.parseShort(value);
-        }
-        return null;
-    }
-
-    public boolean isAnnotationDefaultLiteral()
-    {
-        return m_annotations.get("default_literal") != null;
-    }
-
-    public String getAnnotationValue()
-    {
-        Annotation ann = m_annotations.get("value");
-        if (ann != null)
-        {
-            return ann.getValue();
-        }
-        return null;
-    }
-
-    public Short getAnnotationPosition()
-    {
-        Annotation ann = m_annotations.get("position");
-        if (ann != null)
-        {
-            String value = ann.getValue();
-            if (value.equals("-1"))
-            {
-                return null;
-            }
-            return Short.parseShort(value);
-        }
-        return null;
-    }
-
-    public boolean isAnnotationDefault()
-    {
-        return m_annotations.get("default") != null;
-    }
-
-    public String getAnnotationDefaultValue()
-    {
-        Annotation ann = m_annotations.get("default");
-        if (ann != null)
-        {
-            return ann.getValue();
-        }
-        return "";
-    }
-
     public boolean isIsPlain()
     {
         if (m_typecode != null && !isAnnotationOptional())
@@ -223,40 +80,6 @@ public class Member implements Notebook
             return m_typecode.isIsBounded();
         }
         return false;
-    }
-
-    public boolean isAnnotationId()
-    {
-        return null != m_annotations.get(Annotation.id_str);
-    }
-
-    public String getAnnotationIdValue() throws RuntimeGenerationException
-    {
-        Annotation ann = m_annotations.get(Annotation.id_str);
-        if (ann == null)
-        {
-            throw new RuntimeGenerationException("Error in member " + m_name + ": @" + Annotation.id_str +
-                    " annotation not found.");
-        }
-
-        return ann.getValue();
-    }
-
-    public boolean isAnnotationHashid()
-    {
-        return null != m_annotations.get(Annotation.hashid_str);
-    }
-
-    public String getAnnotationHashidValue() throws RuntimeGenerationException
-    {
-        Annotation ann = m_annotations.get(Annotation.hashid_str);
-        if (ann == null)
-        {
-            throw new RuntimeGenerationException("Error in member " + m_name + ": @" + Annotation.hashid_str +
-                    " annotation not found.");
-        }
-
-        return ann.getValue();
     }
 
     /*!
@@ -303,8 +126,6 @@ public class Member implements Notebook
     private String m_name = null;
 
     private TypeCode m_typecode = null;
-
-    private HashMap<String, Annotation> m_annotations = null;
 
     public static final int MEMBER_ID_INVALID = 0x0FFFFFFF;
